requirements.txt


Django==3.2.18
djongo==1.3.6

h5py==3.1.0

keras==2.6.0
keras-nightly==2.5.0.dev2021032900

numpy==1.19.5

opencv-python==4.7.0.72

packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0

python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1

tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
typing-extensions==3.7.4.3
